requirements.txt
pypdf==5.2.0
pyttsx3==2.98
